+++ packages/client-mobile/src/components/Map/MapCamera/utils.ts
